fix for merging into 10.5

+++ spd_db_mysql.cc
@@ -2530,10 +2530,10 @@ int spider_db_mbase::print_warnings(
             DBUG_PRINT("info",("spider row[0]=%s", row[0]));
             DBUG_PRINT("info",("spider row[1]=%s", row[1]));
             DBUG_PRINT("info",("spider row[2]=%s", row[2]));
-            int res_num =
-              (int) my_strtoll10(row[1], (char**) NULL, &error_num);
-            my_printf_error(res_num, row[2], MYF(0));
-            error_num = res_num;
+            longlong res_num =
+              (longlong) my_strtoll10(row[1], (char**) NULL, &error_num);
+            my_printf_error((int) res_num, row[2], MYF(0));
+            error_num = (int) res_num;
             row = mysql_fetch_row(res);
           }
         }
